pyproject.toml
[project]
name = "skillbridge-assignment"
version = "0.1.0"
description = ""
authors = [
    {name = "Ayush Sharma",email = "sharmaayush9118@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.45.0,<0.46.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "jose (>=1.0.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pydantic[email] (>=2.13.3,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]
[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
